improve create page, rename create, add document title
